Tidy: split TrustTier into a separate file
TurstTier implementation is sufficiently complex to be in its own file.
This makes the top-level EAR structure easier to find inside ear.go.
